requirements.txt
click==8.1.7
cryptography==41.0.5
curl-cffi==0.5.10
Cython==3.0.6
numpy==1.26.2
onnxruntime==1.16.3
opencv-python==4.8.1.78
Pillow==10.1.0
pydantic==2.5.2
PyExecJS==1.5.1
pyppeteer==1.0.2
scipy==1.11.4
tqdm==4.66.1